Move fetchProducts to a dedicated file
